+++ src/Cdn/Cdn/Common/AzureCdnCmdletBase.cs
@@ -14,6 +14,9 @@
 
 using System;
 using System.Collections.Generic;
+using System.Globalization;
+
+using Microsoft.Azure.Commands.Cdn.Helpers;
 using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
 using Microsoft.Azure.Commands.ResourceManager.Common;
@@ -25,7 +28,6 @@ namespace Microsoft.Azure.Commands.Cdn.Common
     /// <summary>
     /// Base class of Azure Cdn Cmdlet.
     /// </summary>
-    [GenericBreakingChange("When you bump up Az.CDN module version, there will be changes to parameters and output, please use the new version", "2.0.0")]
     public class AzureCdnCmdletBase : AzureRMCmdlet
     {
         private ICdnManagementClient _cdnManagementClient;
@@ -78,6 +80,14 @@ namespace Microsoft.Azure.Commands.Cdn.Common
             {
                 action();
             }
+        }
+
+        protected override void BeginProcessing()
+        {
+            CdnMessageHelper.WriteMessage(this,
+                $"Breaking changes in the cmdlet: { MyInvocation.InvocationName}" + Environment.NewLine +
+                "- When you bump up Az.CDN module version, there will be changes to parameters and output, please use the new version: 2.0.0");
+            base.BeginProcessing();
         }
     }
 }